fixing scheduling algorithm
int -> long long
fixing score for ratio

+++ Scheduling/Scheduling.h
@@ -9,12 +9,12 @@ struct Job
 	Job(int weight, int length, bool dif)
 		:	weight(weight)
 		,	length(length)
-		,	score(dif ? weight - length : weight / length)
+		,	score(dif ? weight - length : (double)weight / (double)length)
 	{}
 
-	int weight;
-	int length;
-	int score;
+	int		weight;
+	int		length;
+	double	score;
 };
 
 class Scheduling
@@ -23,11 +23,11 @@ public:
 	Scheduling (bool dif=true);
 	~Scheduling();
 
-	int getCompletionTime();
+	long long getCompletionTime();
 
 private:
-	int getCompletionTimeDifference	();
-	int getCompletionTimeRatio		();
+	long long getCompletionTimeDifference	();
+	long long getCompletionTimeRatio		();
 
 	void jobsFromFile();
